==> notebooks/01_PL_02_Boeing_network.ipynb <==
import geopandas as gpd

import os
import sys

module_path = os.path.abspath(os.path.join('../'))
if module_path not in sys.path:
    sys.path.append(module_path)
import src.analysis as an

# Medellin

nodes = gpd.read_file('../data/input/shape/network_boeing/nodes/medellin-561_nodes.shp')
edges = gpd.read_file('../data/input/shape/network_boeing/edges/medellin-561_edges.shp')

zo = ['floresta_HH','moravia_LH','aguacatala_HL']
z = zo[0]

for z in zo:
    zone_blocks = gpd.read_file(f'../data/output/shape/zone_blocks/{z}/{z}_blocks.shp')
    zone_blocks = zone_blocks.to_crs('32618')
    boeing_nodes, boeing_edges = an.get_boeing_network(zone_blocks, nodes, edges, 500)
    boeing_nodes.to_file(f'../data/output/shape/network_boeing/{z}/{z}_boeing_nodes.shp')
    boeing_edges.to_file(f'../data/output/shape/network_boeing/{z}/{z}_boeing_edges.shp')

# Guadalajara

nodes = gpd.read_file('../data/input/shape/network_boeing/nodes/guadalajara-88_nodes.shp')
edges = gpd.read_file('../data/input/shape/network_boeing/edges/guadalajara-88_edges.shp')

zo = ['miramar_LH','colinas_HL','providencia_HH']
z = zo[0]

for z in zo:
    zone_blocks = gpd.read_file(f'../data/output/shape/zone_blocks/{z}/{z}_blocks.shp')
    zone_blocks = zone_blocks.to_crs('32613')
    boeing_nodes, boeing_edges = an.get_boeing_network(zone_blocks,nodes,edges,500)
    boeing_nodes.to_file(f'../data/output/shape/network_boeing/{z}/{z}_boeing_nodes.shp')
    boeing_edges.to_file(f'../data/output/shape/network_boeing/{z}/{z}_boeing_edges.shp')
