==> notebooks/01_PL_03_Tessellations_network.ipynb <==
# 01_PL_03_Tessellations_network

This notebook creates a __network from block tesselations__.

It uses the most updated version of network_from_tessellation() in order to __include the 'consolidate' argument__, which triggers the use of function ox.consolidate_intersections() function with a given tolerance.

## Import libraries

first_folder_path = "../"

import geopandas as gpd
import pandas as pd

import os
import sys
module_path = os.path.abspath(os.path.join(first_folder_path))
if module_path not in sys.path:
    sys.path.append(module_path)
    import src
else:
    import src

## Main function

def main(gdf_input_path, crs, gdf_output_path):
    
    # Load the blocks shapefile
    print("--- Loading blocks shapefile...")
    gdf = gpd.read_file(gdf_input_path)
    
    # Create the tessellations network
    print("--- Creating tessellations network...")
    nodes, edges = src.network_from_tessellation(gdf, crs, consolidate=consolidate)
    
    # Save the tessellations network locally
    print("--- Saving tessellations network...")
    if consolidate[1]==0: 
        nodes.to_file(gdf_output_path+f"{city}_tessellations_nodes_before_consolidation.shp")
        edges.to_file(gdf_output_path+f"{city}_tessellations_edges_before_consolidation.shp")
    else:
        nodes.to_file(gdf_output_path+f"{city}_tessellations_nodes_consolidation_{consolidate[1]}m.shp")
        edges.to_file(gdf_output_path+f"{city}_tessellations_edges_consolidation_{consolidate[1]}m.shp")

## Run main function

# Define city ('guadalajara' or 'medellin')
city = 'medellin'

# Consolidate the tessellations network? (True/False, consolidation_tolerance) [Check function ox.consolidate_intersections()]
# Project consolidations used for the project:
# False, 0 meters (Used to combine the boeing and tessellations network into the project network)
# True, 5 meters (Used to calculate intersections density as a physical variable in notebook 02_PV_03 and 02_PV_07)
consolidate = (True,5)

# Variables are adjusted according to city.
if city == 'guadalajara':
    # CRS (projected)
    crs = "EPSG:32613"
    # Directory where city's blocks shapefile is stored
    gdf_input_path = first_folder_path + "data/input/shape/Manzanas_Guadalajara/mza2020_guadalajara_32613.gpkg" 
    # Directory (folder) where the resulting tessellations network will be stored
    gdf_output_path = first_folder_path + "data/output/shape/network_tessellations/guadalajara/"

elif city == 'medellin':
    # CRS (projected)
    crs = "EPSG:32618"
    # Directory where city's blocks shapefile is stored 
    gdf_input_path = first_folder_path + "data/input/shape/Medellin_blocks_DANE_2018_new/Medellin_blocks_DANE_2018_new.shp" 
    # Directory (folder) where the resulting tessellations network will be stored
    gdf_output_path = first_folder_path + "data/output/shape/network_tessellations/medellin/"

# Run main function
main(gdf_input_path, crs, gdf_output_path)
