==> notebooks/01_PL_07_n_survey.ipynb <==
from scipy.stats import norm
import pandas as pd
import numpy as np

# Análisis encuesta OD

# Importar la encuesta Origen-Destino del ANVA de 2017
OD = pd.read_csv('../data/input/table/EOD_2017.csv',sep=';',low_memory=False)
OD.head()

# Numero total de registros
n_viajes = len(OD)
# Numero total de personas en base al ID
n_personas = len(OD['ID_CONCAT'].unique())
# Filtrar las personas que caminana
n_personas_caminando = len(OD[OD['DESC_M_Principal'] == 'Caminata ']['ID_CONCAT'].unique())
# Porcentaje de personas que caminan
p_personas_caminando = n_personas_caminando/n_personas
print(f'Porcentaje de personas caminando en todo Medellín: {np.round(p_personas_caminando,4)}')

# Seleccion manual de las zonas SIT de cada territorio de analisis
SIT_O_options_floresta = [195, 193, 200, 194, 201]
SIT_O_options_moravia = [68, 69, 70, 64, 63]
SIT_O_options_aguacatala = [256, 260, 257, 258, 261]
SIT_O_options = SIT_O_options_floresta + SIT_O_options_moravia + SIT_O_options_aguacatala

# Extraer la información de la encuesta OD solamente para las zonas SIT de cada territorio
OD_floresta = OD[OD['SIT_O'].isin(SIT_O_options_floresta)]
OD_moravia = OD[OD['SIT_O'].isin(SIT_O_options_moravia)]
OD_aguacatala = OD[OD['SIT_O'].isin(SIT_O_options_aguacatala)]
OD_TAs = OD[OD['SIT_O'].isin(SIT_O_options)]

# Floresta
n_viajes_floresta = len(OD_floresta)
n_personas_floresta = len(OD_floresta['ID_CONCAT'].unique())
n_personas_caminando_floresta = len(OD_floresta[OD_floresta['DESC_M_Principal'] == 'Caminata ']['ID_CONCAT'].unique())
p_personas_caminando_floresta = n_personas_caminando_floresta/n_personas_floresta

# Moravia
n_viajes_moravia = len(OD_moravia)
n_personas_moravia = len(OD_moravia['ID_CONCAT'].unique())
n_personas_caminando_moravia = len(OD_moravia[OD_moravia['DESC_M_Principal'] == 'Caminata ']['ID_CONCAT'].unique())
p_personas_caminando_moravia = n_personas_caminando_moravia/n_personas_moravia

# Aguacatala
n_viajes_aguacatala = len(OD_aguacatala)
n_personas_aguacatala = len(OD_aguacatala['ID_CONCAT'].unique())
n_personas_caminando_aguacatala = len(OD_aguacatala[OD_aguacatala['DESC_M_Principal'] == 'Caminata ']['ID_CONCAT'].unique())
p_personas_caminando_aguacatala = n_personas_caminando_aguacatala/n_personas_aguacatala

# Medellin
n_viajes_TAs = len(OD_TAs)
n_personas_TAs = len(OD_TAs['ID_CONCAT'].unique())
n_personas_caminando_TAs = len(OD_TAs[OD_TAs['DESC_M_Principal'] == 'Caminata ']['ID_CONCAT'].unique())
p_personas_caminando_TAs = n_personas_caminando_TAs/n_personas_TAs

print('Porcentaje de personas caminando en todo Medellín\t',np.round(p_personas_caminando,4))
print('Porcentaje de personas caminando en todo Floresta\t',np.round(p_personas_caminando_floresta,4))
print('Porcentaje de personas caminando en todo Moravia\t',np.round(p_personas_caminando_moravia,4))
print('Porcentaje de personas caminando en todo Aguacatala\t',np.round(p_personas_caminando_aguacatala,4))
print('Porcentaje de personas caminando en los 3 TA\t\t',np.round(p_personas_caminando_TAs,4))

# n encuestas

# https://online.stat.psu.edu/stat506/lesson/2/2.3

error = 0.05
confidence = 90
alpha = (100-confidence)/100
# p = p_personas_caminando_TAs
p = 0.5
z = norm.ppf(alpha/2)
z

# Formula general
# (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))

# Floresta
N = 23158
n_floresta = (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))

# Moravia
N = 32863
n_moravia = (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))

# Aguacatala
N = 15565
n_aguacatala = (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))

print(f'Considerando p = {np.round(p,4)}')
print('Encuestas necesarias en Floresta\t',int(np.round(n_floresta,0)))
print('Encuestas necesarias en Moravia\t\t',int(np.round(n_moravia,0)))
print('Encuestas necesarias en Aguacatala\t',int(np.round(n_aguacatala,0)))
print('Total',int(np.round(n_floresta,0))+int(np.round(n_moravia,0))+int(np.round(n_aguacatala,0)))

print('Considerando el porcentaje de personas que caminan en cada TA')
p = p_personas_caminando_floresta
N = 23158
print('Encuestas necesarias en Floresta\t',int(np.round((N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p)),0)))
p = p_personas_caminando_moravia
N = 32863
print('Encuestas necesarias en Moravia\t\t',int(np.round((N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p)),0)))
p = p_personas_caminando_aguacatala
N = 15565
print('Encuestas necesarias en Aguacatala\t',int(np.round((N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p)),0)))
print('Total',187+224+66)

# Metodología

Para conseguir el número de encuestas se utilizó la fórmula de este [enlace](https://online.stat.psu.edu/stat506/lesson/2/2.3) donde se nos dice que el tamaño de muestra será:
$$
n = \dfrac{N\cdot p\cdot (1-p)}{(N-1)\dfrac{d^2}{z^2_{\alpha/2}} + p\cdot (1-p)}
$$
Donde:
- $N$ es el tamaño poblacional.
- $d$ es el error esperado en las encuestas.
- $z$ es la significancia, hallada como 1 menos el nivel de confianza que se tenga.
- $p$ es la proporción que se quiere estimar.

Nuestra información será la siguiente: Conocemos el tamaño poblacional de cada uno de nuestros Territorios de Análisis (TA) al saber las personas que viven en las manzanas que pertenecen a cada TA con la información del censo del DANE, consideraremos un error esperado del 5% y trabajremos con un nivel de confianza del 90%. 

Para el parámetro $p$ se suele utilizar un valor de 0.5, el cuál representa el caso más desfavorable en donde no se tiene información detallado de lo que se está estimando. Para nuestro caso de estudio, nosotros estamos interesados en conocer como son las condiciones de caminabilidad de los TAs. Por lo tanto, nos interesará saber la opinión de las personas residentes en nuestros TA que se desplazan caminando. Para esto utilizaremos la información de la encuesta Origen-Destino del Área Metropolitana del Valle de Aburrá (AMVA) del 2017, en donde se registran viajes representativos que realizan las personas en todo le territorio.

**Opción 1:** De la encuesta podemos saber la cantidad de personas que realizan viajes caminando en cada uno de nuestros TA, y por consiguiente podemos saber el porcentaje de la población que realiza sus viajes caminando en cada uno de nuestros TAs. Este porcentaje de personas caminando será el valor que utilizaremos como parámetro $p$.

# Guadalajara

# https://online.stat.psu.edu/stat506/lesson/2/2.3

error = 0.05
confidence = 90
alpha = (100-confidence)/100
p = 0.269
z = norm.ppf(alpha/2)
z

# Miramar
N = 11395
n = (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))
int(np.round(n,0))

# Colinas de San Javier
N = 3401
n = (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))
int(np.round(n,0))

# Providencia
N = 3752
n = (N*p*(1-p))/((N-1)*error**2/z**2 + p*(1-p))
int(np.round(n,0))
